verification/z3_verifier.py
"""Z3 形式化验证器"""
from typing import Dict, Any, List
from z3 import Solver, Int, sat, unsat, And


class Z3Verifier:
    """Z3 形式化验证器"""

    def __init__(self, timeout: int = 30):
        self.timeout = timeout

    def verify(self, code: str, constraints: List[str]) -> Dict[str, Any]:
        """
        验证代码安全性

        Args:
            code: 待验证代码
            constraints: 安全约束列表

        Returns:
            验证结果
        """
        # 1. 提取代码逻辑（简化版）
        formula = self.extract_logic_formula(code)

        # 2. 定义安全规则
        rules = self.define_security_rules(constraints)

        # 3. 创建求解器
        solver = Solver()
        solver.set("timeout", self.timeout * 1000)

        # 4. 注入逻辑公式
        solver.add(formula)

        # 5. 注入安全规则
        for rule in rules:
            solver.add(rule)

        # 6. 执行验证
        result = solver.check()

        # 7. 生成证明
        if result == unsat:
            proof = {
                "status": "verified",
                "result": "unsat",
                "formula": str(formula),
                "rules": constraints
            }
        else:
            model = solver.model()
            proof = {
                "status": "failed",
                "result": "sat",
                "counterexample": str(model),
                "rules": constraints
            }

        return proof

    def extract_logic_formula(self, code: str):
        """
        提取代码逻辑公式

简化版实现，实际应用中需要 Tree-sitter 解析
        """
        # 模拟提取逻辑公式
        # 这里应该使用 Tree-sitter 解析 AST 并转换为数学公式
        path_var = Int('path')
        return path_var

    def define_security_rules(self, constraints: List[str]):
        """
        定义安全规则

将约束转换为 Z3 表达式
        """
        rules = []
        for constraint in constraints:
            # 简单的规则转换
            if "禁止访问" in constraint:
                # 提取禁止路径
                import re
                paths = re.findall(r'([^\s]+)', constraint)
                if len(paths) > 0:
                    path_var = Int('path')
                    forbidden = [Int(f"idx_{i}") for i in range(len(paths))]
                    rule = path_var not in forbidden
                    rules.append(rule)
        return rules
