requirements.txt
# AetherHub 依赖列表

# 核心依赖
z3-solver==4.12.6
tree-sitter==0.21.0
wasmtime==16.0.0
python-dotenv==1.0.0

# 开发依赖
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# 可选依赖
# openai==1.3.0  # 如果需要使用 OpenAI API
